refactor context and themeSwitcher
