+++ src/js/pages/ChannelsList.jsx
@@ -1,7 +1,6 @@
 import React from "react"
 import ChannelCard from "../components/ChannelCard"
 import Settings from "../components/Settings"
-import ThemeSwitcher from "../components/ThemeSwitcher"
 
 export default function ChannelsList({ channels }) {
     const cards = []
@@ -15,7 +14,6 @@ export default function ChannelsList({ channels }) {
 
     return (
         <>
-            <ThemeSwitcher />
             <Settings />
             <div className="wrapper">
                 <div className="radio-title">
